Turn off the strict aliasing warnings in the autoconf gcc build.
Details:

This is slightly dangerous because it can flag real problems the
optimizer will trip over. However, Ghostscript's method of implementing
inheritance through macro expansion in structure definitions generates
hundreds of (not problematic) warnings because the compiler doesn't
actually trace through to verify that the referenced members are at the
same offset.

We cannot fix this easily without making the code much more convoluted,
so in the interests of noticing and fixing other warnings, we've decided
to turn off this particular check.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7671 a1074d23-0009-0410-80fe-cf8c14f379e6