Rewrite the test script to use a custom test class.
Previously we used the python unittest module, but we don't use most of 
its features, and it's quite weak for this kind of file-based external 
testing. So we lose on using a familiar api, but this doesn't add much 
code, and we can now add long-desired features like 'xfail' results.

This grew out of work revising the Ghostscript test code for 
parallelization.


git-svn-id: http://svn.ghostscript.com/jbig2dec/trunk@452 ded80894-8fb9-0310-811b-c03f3676ab4d