documents ownership of modules.  For now this resides here instead of gs/doc because some of the devices are not part of gs.  We will be renaming the ghostpcl repository to ghostpdl in the future at which time the location of this file will make more sense.  Changes to this file that involve relinquishing or trading a module should be agreed upon by affected parties.  This file is maintained in a public area and should not contain confidential information.
Known problems:

Devices are missing, in particular any device that is not normally
compiled with gs, language switch, or pcl with gcc is likely to be
missing.

Ownership of other categories: testing, configuration management, new
technologies (xps) need to be added.

Large customer "ownership" is missing.  We might be able to add these
by customer number since this is in a public area.

PS This checkin is a test of committing with svk.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@2759 06663e23-700e-0410-b217-a244a6096597