Fix (type 1 hinter) : An improvement for very small matrices.
DETAILS :

Debugged with this :

/Courier findfont 0.000000000000000000000000000000001 scalefont setfont
72 72 moveto
(ASDFGGH) show
showpage

In fraction_matrix__set the old code computed this->bitshift = 137,
and then

	this->denominator = 1 << this->bitshift

gave a nonsnse result 512.
Here is the assembler code for the latter statement built with MSVC8 :

	mov         ecx,dword ptr [eax+14h]
	shl         edx,cl

It simply drops higher bytes.
Looks as MSVC compiler bug.

The new code replaces it with strong zero.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7702 a1074d23-0009-0410-80fe-cf8c14f379e6