Non-backwards compatible change: use consistent reference count allocation discipline for color spaces. Previously, color spaces were allocated on stack, on the heap, as sub-structures within other color space structures, and were very commonly copied. Reference counting was used for the "params" block of some, but not all, color spaces, but there were correctness problems with that logic.
In addition, a number of "const" qualifiers have been removed, because
the action of bumping the reference count on an object is not consistent
with the C semantics of const.

This change is required to resolve problems with enhancement 688924. It
should also help with memory leaks.

Other clients, such as the PCL interpreter, will require changes to be
compatible. For simple device colorspaces, the change to gs_setcmykcolor()
(gscolor1.c) is representative of the code change required.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7765 a1074d23-0009-0410-80fe-cf8c14f379e6