XPS build fixes.
Declare the expat link dependency through xps.dev instead of EXTRALIBS.
The old mechanism wasn't working for me, so there may be another bug
here.

Likewise include the gstype2 implementation through a dependency in
xps.dev. I'm not sure if this is correct, but if it is, all the other
required modules should also be declared. FEATURE_DEVS is for optional
features.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@2802 06663e23-700e-0410-b217-a244a6096597