Implementing the PS operator .sort in C language.
DETAILS :

This is a partial fix for bug 688971
"huge performace problem (with large TT font?)".
Patch from SaGS.

The old PS implementation of lib\gs_init.ps::.sort uses a
slow O(n^2) algorithm. Plus, array indexing operations are lenghty
in PostScript. It seems to me this was initially written to sort
the very few "%disk*%" names, so speed was irrelevant. Now, it is
used to sort 'loca' tables (in some damaged TTFs).

The patch implements the heap sort algorithm from the programmer's
bible. This algorithm is guaranteed to be O(n log n) in the worst
case. Variable's names are those from Knuth's book. Labels denote the
algorithm's steps.

The implementation retains the maximum generality for the predicate,
i.e. it can be anything that is able to compare 2 objects on the
o-stack and leave a bool in their place. The predicate is not
restricted to what's available to a FunctionType 4 function.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7868 a1074d23-0009-0410-80fe-cf8c14f379e6