Implementation of a new non-linear image filter that prevents dropout when images are scaled down to a 1-bit per component device. Bugs 689147 for customer 531 and image quality problem for customer 780. This filter is used whenever /Interpolate == true (or -dDOINTERPOLATE command line option is used) and the image is being scaled down on a 1-bit per component device.
Interpolation throughput of gray or cmyk images is also improved by
a factor of about 2:1 (depending on the data content) even when
the standard Mitchell filter is used. The performance improvement
when using this new filter is even higher since the filter itself
uses less CPU time.

DETAILS:

This filter uses the darkest of the image pixels that are within a
device pixel, rather than the default behaviour which uses the PS
'center of pixel' rule (i.e., whichever image pixel covers the
center of the device pixel is used to color the device pixel).

The center of pixel rule can result in missing narrow black
lines or single dots in patterns when a source image is scaled down.

One side effect of this filter is that narrow white features or
small white areas in patterns can be lost, effectively darkening
the result. For now the customers needing this are not concerned
with small white-on-black areas becoming solid black. Many real
world printers tend to do this anyway due to 'dot gain'.

If we encounter the need for preserving small white areas, we
can change the filter somewhat to prevent this effect.

EXPECTED DIFFERENCES:

None. (except improved performance for Interpolated images).

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7886 a1074d23-0009-0410-80fe-cf8c14f379e6