Fix (shadings) : Early skip parts that fall outside the clipping box.
DETAILS :

This improves performance for the test case of
the bug 689189 "PDF fails with /unregistered in --shfill--".

The patch implements a bbox intersection code pattern to
several decomposition functions. It speeds up those shadings,
in which the shading area is significantly larger
than the clipping box.

Rather it speeds up the test case,
it needs a further improvement for shading functions.
A small noize (being lesser than shading smoothness) to be ignored
while a function monotonity check.

Minor changes :
1. In some functions the old bbox intersection code
   becomes more uniform.
2. bbox_of_points moved up.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7892 a1074d23-0009-0410-80fe-cf8c14f379e6