Fix (shadings) : Implement a color index cache.
DETAILS :

This partially improves performance for the test case of
the bug 689189 "PDF fails with /unregistered in --shfill--".

This change is algorithmically equivalent.
The old code redundantly converts client colors
to device colors. This patch implements a cache for
the conversion in the new module gscicach.h .
It works for contone devices only.

For the test case it reduces the number of color conversions
in 20+ times. However a rocket speeding up is not obtained.

This patch is an experimental one.
We have got an alternative approach,
which is more complicated but works for any device.
The alternative is to pass converted colors through the decomposition recursion.
We commit this now mainly for archiving purpose.
To disable it replace the call to gs_color_index_cache_create with NULL.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7908 a1074d23-0009-0410-80fe-cf8c14f379e6