Add support for /UserUnit. Also fix some related bugs that stayed in the way while testing it. Patch from SaGS. Bug 688124 from customet 870, bug 688359.
DIFFERENCES:
None

DETAILS:
(A) The basic implementation: scaling the PS user space

/UserUnit implementation is very similar to -dPDFFitPages.

- If -dPDFFitPages=true, then /UserUnit is ignored; no matter
  how small or haw large is 1 PDF user space unit, it has to
  be scaled so that the PDF page fits on the paper.
- Else, the /UserUnit scales the PS user space, exactly like
  -dPDFFitPage does. This solves [almost] everything related
  to drawing marks on the page.

This also means that all problems with -dPDFFitPages affect
the implementation of /UserUnit. The biggest problem was to
get rid of those, especially for PDF->PDF "conversion" where
some elements of the source PDF (outlines, links...) needed
to be preserved.

(B) Don't scale the border width in a border style dict
    (the change to pdf_draw.ps)

See comment in code. Note that a border width in a border style
ARRAY is specified in user space units, so it grows with
/UserUnit and the scaling of the PS user space suffices.

(C) No more /PAGES for /CropBox
    (pdf_main.ps hunk @@ -129,10 +129,6 @@
    and the "pget" instead of "knownoget" for /PAGE pdfmark)

/UserUnit, /Rotate and the translation due to non-(0,0) PDF page
origin are "flattened" into the page; also -dPDFFitPage scales
the page. This means the even if 2 source PDF pages had the same
/CropBox, in the destination PDF these may need different
/CropBox-es.

Example:

source:  page #1 /UserUnit 1 and /CropBox [0 0 100 100]
	 page #2 /UserUnit 2 and /CropBox [0 0 100 100] (the same)
becomes: page #1 no /UserUnit and /CropBox [0 0 100 100]
	 page #2 no /UserUnit and /CropBox [0 0 200 200] (differs)

so the /CropBox cannot be inherited anymore. The new code puts a
/CropBox into each page that has or inherits one.

NOTE:
  There are 2 more places in pdf_main.ps that do a "knownget"
  for /CropBox, one a few lines after "%****** DOESN'T HANDLE
  COLOR TRANSFER YET ******" and one after "(Adobe Tech Note
  5407, sec 9.2)". I think both should be using "pget".

(D) EXTRA: /CropBox-es in intermediate /Pages were ignored

Old code preserved only /CropBox-es that appeared in the root
PDF /Pages object and in /Page objects, the ones in intermediate
nodes of the /Pages tree being lost. In the new code, this
gets fixed for PDF->PDF as a side effect of the implementation
for (C). See also the note above for clipping of the marks
drawn on the page.

(E) PDF->PS "default" user space transform

The new code computes a matrix that transforms the PDF default
user space of a page to the PS default user space. This matrix
accounts for the rotation (/Rotate), scaling (-dPDFFitPage or
/UserUnit) and any translation needed to move the PDF
lower-left corner the the lower-left corner of the paper.

- This is done in the new pdf_main.ps::pdf_PDF2PS_matrix,
  which inherits, with the needed changes, almost all of the
  code in the old .pdfshowpage_Install.

- The matrix is page-specific because different pages may have
  different dimensions (so -dPDFFitPage scales them
  differently), different /Rotate or /UserUnit.

- pdf_main.ps::pdf_cached_PDF2PS_matrix is a utility proc that
  ensures the matrix for a given page is computed, caches it
  in the PDF page dictionary under the key given by
  pdf_main.ps::PDF2PS_matrix_key, then returns the matrix.

- (The definition for PDF2PS_matrix_key exists only to allow
  binding of a complicated name into pdf_cached_PDF2PS_matrix;
  avoids doing "(complicated.name) cvn" at run-time.)

- This matrix is currently used:
  (E.1) by .pdfshowpage_Install (which is now reduced to 2
	lines) for setting up the PS user space;
  (E.2) for transforming the /Crop- or /MediaBox in
	pdfshowpage_setpage
  (E.3) to transform coordinates in view destinations (used
	by outline entries, PDF links ...).

(F) /Orientation now always 0

pdfmark does not work correctly with /Orientation != 0 (long
story). The old code used the /Orientation page device
parameter to handle the /Rotate from PDF pages. The new
code always sets /Orientation to 0 and handles /Rotate by
explicitely doing a "rotate" (in pdf_PDF2PS_matrix).

- Avoids GS-specific hackery otherwise needed to work around
  pdfmark problems when /Orientation != 0.

- Simplifies the code, because a single transformation matrix
  needs to be computed both for setting up the PS user space
  and for transforming varions coordinates used in pdfmarks
  (/CropBox, view destinations).

(G) PDF->PDF-migrated view destinations were wrong
    (pdf_main.ps hunks @@ -939,6 +935,45 @@
    and @@ -947,18 +982,30 @@)

Coordinates appearing in view destinations need to be recomputed,
due the PS default user space, AS USED BY the pdfwrite driver,
not being identical to the original PDF default user space. The
list of causes included rotation due to /Rotate (implemented either
with /Orientation or a simple "rotate"), translation due to
non-(0,0) PDF page origin, and scaling due to -dPDFFitPages; now
we add /UserUnit to this list.

(H) EXTRA: -dPDFFitPage now chooses portrait or landscape
    (pdf_main.ps, near the end of hunk @@ -1031,62 +1078,133 @@)

If -dPDFFitPage, the code after "% Preserve page size," chooses
portrait or landscape orientation depending on the PDF page's
width:height ratio. I consider this results in a better "fit to
page". Example: PDF with mixed portrait + landscape letter
pages, to be printed on A4 paper. Old code sometimes fitted
landscape pages on portrait paper.

(I) EXTRA: better placement of imaged area

If -dPDFFitPage and the PDF page's width:height ratio differs
from the paper's width:height, some unused space remains. With
the old code, this extra space was placed at left/right/top/bottom
depending on the page's /Rotate. New code always puts the extra
space either at right or top (depending only on the PDF page
being relatively "taller" or "wider" than the paper). This
is mainly a side effect of not using /Orientation anymore.

(J) -dNoUserUnit

I added a new option that can be used to disable processing of
/UserUnit. Named NoUserUnit, defaults to "false" meaning
/UserUnit being taken into account. I implemented this default
following Adobe Reader's 7.0.7/Windows default, but see note.

Note:
  I suggest to set the default to ignore UserUnit, and have a
  -dDoUserUnit option to activate it. I can do this change
  if desired.
  I'll explain the reson for such a choice through an example:
- I THINK that UserUnit was introduced by Adobe as part of
  it [Adobe] entering the CAD world.
- Consider a floorplan plotted on a sheet of paper at a certain
  scale, let's say 1:50.
- In this scenario, the PDF page corresponds to the plotted
  paper, so it has a MediaBox of that size.
- If the scale is 1:50, set UserUnit = 50. This allows someone,
  given a suitable UI, to easily and accurately MEASURE
  various elements of the floorplan.
- Ghostscript does not have such a UI, and I think such a UI
  is beyond GS's purpose.
- GS's role, however, is to PRINT that PDF in order to obtain
  the "plotted" floorplan.
- To obtain the equivalent of the plotted paper, printing must
  ignore UserUnit. Observing UserUnit for printing would
  require a building-sized sheet of paper!

(K) TESTING DETAIL: "transform" returns reals

"<x> <y> <matrix> transform" returns 2 realtype objects, even
when <x> <y> are integertype and the matrix is [1 0 0 1 0 0]
(identity, containing only integers). This makes, for example,
a Media- or CropBox of [0 0 612 792] in a source PDF to be
written as [0 0 612.0 792] in the destination, which is
annoying if comparing the output of unpatched and patched
Ghostscript.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@7999 a1074d23-0009-0410-80fe-cf8c14f379e6