Fix (shadings) : Optimize filling a radial shading extension.
DETAILS :

This partially improves performance for the test case of
the bug 689189 "PDF fails with /unregistered in --shfill--".

The test case includes multiple radial shadings,
which paint an extansion only.
This patch optimizes the case when the paint area is
strongly inside the radial shading extension.
In this case the shading rasterization is replaced with
painting a pure color.

1. The new function R_fill_rect_with_const_color is the main worker.
2. The old patch_color_to_device_color function is renamed into
   patch_color_to_device_color_inline.
3. patch_color_to_device_color is made public for R_fill_rect_with_const_color.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8006 a1074d23-0009-0410-80fe-cf8c14f379e6