Primarily fix for wtsimdi problem with wts dithering phase shift at band boundaries (add band_offset_x, band_offset_y), adding band_offset_x and band_offset_y to the device structure and corresponding initializer changes. Also clean up gdevwts.c MSVC warnings, and add a fast check to gdevbit.c for /dev/null (mirroring a much older change for Windows "nul:" device) OutputFile that aids in benchmarking. Fixes bug #689245 for customer #951.
DETAILS:

Running with the wtsimdi device at default banding now matches the
results with a single band (the wtsimdi device sets BandingAlways,
so this can only be verified if the -dBandHeight= and -dBufferSpace=
options are set large enough for a full page.

EXPECTED DIFFERENCES:

None. (wtsimdi device changes only).

NOTE: 'contrib' devices that do not use the std_device macros to
initialize the device_common part of the device structure will get
a compile error (as the bittagsrgb device did). We need to fix our
bittagsrgb device, and the eps-merge may need changes if there are
devices that do not use the macros (none known at this time).

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8013 a1074d23-0009-0410-80fe-cf8c14f379e6