Fix (shadings) : Optimize path manipulations for shading fill.
DETAILS :

This partially improves performance for the test case of
the bug 689189 "PDF fails with /unregistered in --shfill--".

While filling a shading, the old code
converts the clipping path into a path, then
converts it to a clipping path for intersecting it
with the filling path and with the shading BBox.
These manipulations are neccessary due to the
interface restrictions for gs_shading_fill_path,
which doesn't pass a clipping path.
It causes (1) a bad performance and (2)
a wrong fill adjustment, because the conversion
of a path to a clipping path and back to path
is not identity due to fill adjustment.

The new code is a 1st step for the elimination of gs_shading_fill_path.
This patch excludes this function for raster devices.
Excluding it for higth level devices will be a next step.

The new code moves path manipulation outside the
shading color painting. Now all 3 pathes
(the filling path, the clipping patrh, and the shading BBox)
are processed in gxfill.c . The painting
is done with gs_shading_do_fill_rectangle,
which does not manipulate with pathes.

The patch modifies gx_image_fill_masked_end
to comply with the new inplementation of
gs_shading_do_fill_rectangle when a mask
is converted into a clipping path.

Another optimization avoids the generation
of a filling path for 'shfill'.
Accorgong to PLRM, shfill deals with clipping path,
but it was converted into a path due to
internal Ghostscript constraint for the device interface.
When the device implements fill_path with
the default method gx_default_fill_path,
the conversion to a path is not necessary.

To make use of this fact for optimization, the graphics library
needs to know, whether the output device uses the default
function gx_default_fill_path. This patch introduces a new response
for pattern_manage, which is named 'pattern_manage__shfill_doesnt_need_path'.
The device must respond with 1 if it does not need a path.
The patch provides implementations for raster devices and
for forwarding devices. If some device provides an instandard
fill_path method, the optimization won't work for it
(i.e. it will work as before - will recieve a path for shfill)
unless it provides a proper response to pattern_manage__shfill_doesnt_need_path .

A minor change to gdevpdfb.c fixes a paranoid compiler warning
about a possible use of unubnitialized data.

EXPECTED DIFFERENCES :

The old code erroneusely applied fill adjustment two times
when filling a path with a shading color.
Also fill adjustment was erroneusely applied when
painting a characxter with a shading color.
With new code shadings paint fewer pixels at a shading boundary,
causing a minor difference in rasters :

72dpi :

"442-01.ps"
"446-01-fixed.ps"
"470-01.ps"
"Bug687546.ps"
"Bug689189.pdf"
"chilis_black.pdf"
"chilis_red.pdf"
"Clarke Tate Manns Chinese.ai"
"gradmesh.ai"
"muur.eps"
"S2_Digitalproof-Forum_x3k.pdf"
"shading_prob_800.ps"
"SmoothShading.pdf"
"Testform.v1.0.2.pdf"

300dpi :

"470-01.ps"
"Bug687546.ps"
"Bug689189.pdf"
"chilis_black.pdf"
"chilis_red.pdf"
"Clarke Tate Manns Chinese.ai"
"muur.eps"
"S2_Digitalproof-Forum_x3k.pdf"
"SmoothShading.pdf"

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8017 a1074d23-0009-0410-80fe-cf8c14f379e6