Add a work-around for malloc(0) returning a 0 pointer - call malloc(1) instead. Bug 688532, customer 870.
DETAILS:
This behavior was observer on IBM 4.3 and 5.1. IEEE 1003.1 spec,
which is aligned with ISO C, states:
If the size of the space requested is 0, the behavior is implementation-defined:
the value returned shall be either a null pointer or a unique pointer.

malloc(0) happens when libjasper reads optional tags in the embedded ICC
profile. This doesn't happen too frequently and we can afford to allocate
a few extra bytes.

DIFFERENCES:
None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8039 a1074d23-0009-0410-80fe-cf8c14f379e6