This is a fix for regression bug report 689263.
DETAILS:

Some versions of PhotoShop generate documents that place an extra value on the operand stack and tintTransform replaces it - see bug 549307 for details. Also see the test case of bug 689263. Since GS invokes setcolor as a side-effect of setcolorspace, we must prevent Adobe's hack from pop'ing an extra (significant) item from the operand stack when setcolor is invoked from a setcolorspace context.

The previous patch for bug report 688584, which was committed before I was done with it, was too broad and was undoing Adobe's hack whenever setcolor was invoked (from setcolorspace or otherwise). This change also undoes the previous patch, which was the cause of the error:

Error: /rangecheck in --image--

under the conditions specified in bug report 689263.

EXPECTED DIFFERENCES:

Regression test file Bug688584.ps should now work properly for all output device types.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8048 a1074d23-0009-0410-80fe-cf8c14f379e6