Fix an out of bounds buffer write in the wts_load_halftone logic and fine tune the wts_resolve_one color conversion cache hashing for 4K (the default). Bug 689255 for customer 951.
DETAILS:

Besides fixing the out of bounds logic, I added a 'bufsize' parameter to the
gs_wts_from_buf logic so that we can detect OOB conditions. Note that the
1+hdr_size+cell_size in gx_ht_read_component_wts is retained by this patch
even though it looks bogus (we want to commit the fix so we can send to a
customer, and will look at this later).

In refining the cache stats in the wtsimdi device, the 'fill_empty' case
and the 'collision' case were split to make sure that the hash is decent,
but even 1M cache hits (on ACDsee.prn) only save about 1.5 seconds (out
of 9 seconds). The hash algorithm for smaller collision rates won't really
matter much.

EXPECTED DIFFERENCES:

None. (what did you expect since we don't do regression on wtsimdi yet?)

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8080 a1074d23-0009-0410-80fe-cf8c14f379e6