Fix : Restoring high level objects in clist.
DETAILS :

This patch is a part of the clist expansion problem.
With Henry's test foo4.ps it reduces the clist size in 10+ times.

The clist expansion problem appears to be caused by an old unintentional
change, which was done 3 years ago within the psdcmyk project.
The revision 5056 effectively disabled a big part of the clist language:
all high level objects (text, images, paths) were written to clist
as lots of rectangles. Undoing it now because it is incorrect.
Note that this patch doesn't undo other parts of 5056 change.

With enabling high level objects the psdcmyk device becomes broken.
At least it has problems with text when rendering foo4.ps .
We'll open a separate bug about that.

Rather this fix is a progression, it causes some regressions with
band devices due to other unknown problems.
There are single pixel differences with comparefiles,
which are not visible to users, but they are important
for regression testing. Since the clist size
is a highly important problem now, we commit this fix,
and open a new bug about raster differences.

EXPECTED DIFFERENCES :

The following files have an expected raster differences with band devices,
which may be classified as a regression.
See DETAILS about them.

"034-01.ps"
"289-01.ps"
"Bug688485.pdf"
"Openhuis_pdf_zw.pdf"
"S2_Digitalproof-Forum_x3k.pdf"
"SmoothShading.pdf"
"test.pdf"

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8110 a1074d23-0009-0410-80fe-cf8c14f379e6