Fix (shadings) : Decomposition limit could be too small.
DETAILS :

Bug 689225 "Regression: major differences in 470-01.ps in block 470-03".

This patch improves the tolerance of the shading code to
a wrong device resolution, which can apparently happen
when installing a clipper device. Thus is closes a symptom,
but the bottom of the problem is left unfixed.
We do so due to incoming scheduled release.
We opened a separate bug 689365 about a wronmg HWResolution of clipper device.

EXPECTED DIFFERENCES :

Should fix the recent regression with 470-01.ps .

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8136 a1074d23-0009-0410-80fe-cf8c14f379e6