Fix for seg fault caused by undersized buffer as reported in bug report #687397
DETAILS:

Although the patch submitted by the reporter did appear to fix the problem,
it was more masking it than fixing it. While the analysis of the problem
being that ss->Items was too small was correct, the real cause was a rounding
problem in siscale.c/contrib_pixels(). The problem did not happen on my Mac,
but was reproduced up on peeves because of differnces in the way the two
systems perform rounding.

EXPECTED DIFFERENCES:

None

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8173 a1074d23-0009-0410-80fe-cf8c14f379e6