Fix stdint.h related problems with the Cygwin build (gcc 3.4.4)
DETAILS:

There was a workaround for an obsolete version of Cygwin tools that had
u_int32_t (and other types) instead of uint32_t that is standard on linux
and Cygwin stdint.h now. These changes were needed for the PCL, XPS and
language_switch builds -- the gs build worked thanks to autoconf.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8195 a1074d23-0009-0410-80fe-cf8c14f379e6