Fix (pdfwrite) : Uninitialized defaultWidthXwhile while converting a Type 1 font into Type 2.
DETAILS :

Bug 689406 "Valgrind: Uninitialised defaultWidthX in psf_convert_type1_to_type2()".

Thanks to Alex for the problem localization and for the suggested patch.
This change is equivalent to Alex's patch
expect the better documentation.
See comment in code.

The old code accessed pfont->data.defaultWidthX,
which is never used or initialized elsewhere with a Type 1 font,
because defaultWidthX is a Type 2 feature.

EXPECTED DIFFERENCES :

None.

Actually the probability of the correct result with the old code
is not equal to 1, so some differences with glyph positioning
may appear in some cases. Alex's local test didn't detect a diference
with comparefiles.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8211 a1074d23-0009-0410-80fe-cf8c14f379e6