Fix (font rendering) : alloc_char_bits could fall into infinite loop.
DETAILS :

Bug 689036 "infinite loop in char cache logic".

The character cache algorithm assumes that a pointer to
any allocated cached_char instance appears somewhere in ccache.table .
However from the old experience we know that
this obvious invariant appears frequently broken due
to inaccurate manipulation with cache by clients.
For example, a client sometimes reserves and initializes
a cached_char instance and then fail due to another error
before storing the pointer to ccache.table.

In the old code a cycle in alloc_char_bits
strongly depends on this invariant,
and therefore it is potentially harmful.
In the past we have got a lot of bugs about it.
Now we restrict the cycle with the cache size.
Changed few function prototypes with returning error
if the invariant is broken.
Note that the graphics library interface changes,
so other interpreters must update the calls to
alloc_char_bits.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8219 a1074d23-0009-0410-80fe-cf8c14f379e6