Replace the 'private' define with the now standard C keyword 'static'.
The compilers we support have supported static for some time now, so
it makes sense to use it by default. This also makes the code less
idiomatic, and removes one of the obstacles to compiling Ghostscript
as C++.

The patch also removes support for compilation with -dNOPRIVATE, which
is not longer necessary with modern debuggers.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@2923 06663e23-700e-0410-b217-a244a6096597