Fix (pdf14 device) : Don't allocate transparency buffer for bands that appear outside the transparency bounding box.
DETAILS :

This is an optimization, which should be algorithmically equivalent.

For bands outside the transparency bounding box the old code
allocated a zero length buffer and performed fake operations with it.

The main change is gdevp14.c ln 550.
Others prevent an access to the buffer when it is not allocated.

Note that in some cases the new check is excessive
because (width <= 0 || height <= 0) normally means no buffer.
We check for (buf->data == NULL) anyway to simplify the sefity proof
and documentation.

EXPECTED DIFFERENCES :

None,

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8279 a1074d23-0009-0410-80fe-cf8c14f379e6