Fix (stroking) : Prevent installing an extra clipper device, step 2.
DETAILS :

Bug 689427 "Superposed clippers".

The last patch doesn't fix the problem when stroking
a path with a shading color.

The new code distinguishes 3 cases :

1. When painting happens through low level device methods
which don't accept a clipping path. It happens inside
stroke_fill, so the new code passes the clipper device to it.

2. When the painting happens through fill_path,
which does handle a clipping path. It happens inside
FILL_STROKE_PATH, so the new code doesn't pass the clipper to it,
but passes the clipping path as it worked before the last patch.

3. When the stroking algorithm is called from 'strokepath',
don't need a clipping at all. This case works as before,
but the new code doesn't construct a clipping device.

Minor change : Added a new argument 'dev' to the macro
FILL_STROKE_PATH to simplify the code understanding and debugging.

EXPECTED DIFFERENCES :

None,

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8286 a1074d23-0009-0410-80fe-cf8c14f379e6