Added code to the pxlcolor device (gdevpx.c) to set the color space even if the test for "portrait transformations" fails.
DETAILS:

The pclxl_begin_image() function performs a rather complicated test to
see if the image transformation can be "handled" (line 1434).  If this
test fails the routine falls through to using gx_default_begin_image().
However, in this case the pcl-xl color space is never set and if
the previous color space was eGray the pcl-xl stream will fail with
ExtraData (because 3 times as much data as expected will be written).
Presumably the inverse is also true, but I don't have a test case for it.

The solution is to call pclxl_set_color_space() in the use_default
case (a possibly more correct solution would be to call
pclxl_set_color_palette()).

EXPECTED DIFFERENCES:

None.  The nightly regression doesn't test the pxlcolor device.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8290 a1074d23-0009-0410-80fe-cf8c14f379e6