Remove a spurious CFLAGS instance from the jasper compile line.
DETAILS:

The definition of CC_ passed from the top level makefile should
already include a copy of the CFLAGS for environment overrides,
so the duplicate copy here was unnecessary.

Furthermore, the second expansion of CFLAGS at the end of the
compiler command line could result in picking up a version of
the jasper headers installed elsewhere in the system instead
of the copy in JASSRC, causing the build to fail on undefined
preprocessor symbols, since upstream doesn't have our error
reporting callback.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8301 a1074d23-0009-0410-80fe-cf8c14f379e6