Fix (pdfwrite): Type 1 CharStrings could be converted to invalid type 2 CharStrings if the glyph program used implicit start points for subpaths.
DETAILS:
Bug #689055 "pdfwrite creates a bad Type1C font"

If a type 1 CharString used an implicit point, either the initial point
(0,0) or the point resulting from a closepath, then the type 2 output
code duplicated this. Unfortunately this is illegal in type 2 CharStrings
where each subpath must begin with one of the 'move' instructions.

(gdevpsfx.c) psf_convert_type1_to_type2, create a new boolean variable
'need_moveto' initialised to true. When we encounter a move operation,
set it to false. When we encounter a closepath, set it back to true.
Thus this variable is always true at the beginning of a subpath, and
becomes false as soon as a move instruction is encountered.

While converting type 1 CharStrings, whenever the current instruction is
a line drawing operation, if 'need_moveto' create a move instruction,
taking account of any pending glyph widths.

This is only performed on the second (conversion) pass through the type 1
CharSrtring as we don't need to worry about this during the first (hint
collection) pass.

EXPECTED DIFFERENCES:
None

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8324 a1074d23-0009-0410-80fe-cf8c14f379e6