Fix (pdfwrite): named objects, created via a pdfmark such as /BP, could not have their content modified after the object was completed.
DETAILS:
Bug #689104 "Overwriting transformation /Matrix in XObject previously
created with /BP & /EP pdfmarks".

Some pdfmarks, such as /BP create named objects. These can accumulate
information until a matching pdfmark (/EP in this case) completes the
object. When complete, objects were written to the final destination
PDF file. This made it impossible to later modify the object.

(gdevpfm.c) pdfmark_bind_named_object, change the call to
pdf_substitute_resource to defer writing the object to the
destination file, if the object is named. This means the object is still
available for modification until the end of the job. NB if the object
is not named, we still write it immediately, as there is no mechanism
for modifying an unnamed object, nor do we currently have a mechanism
for maintaining a record of unnamed objects.

EXPECTED DIFFERENCES:
None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8359 a1074d23-0009-0410-80fe-cf8c14f379e6