Add changes for  building COMPILE_INITS=1 without PS (PCL only build). Make sure gs/src/ugcclib.mak does conditional assignment of COMPILE_INITS. Fix main/pcl6_gcc.mak for UFST fco location when COMPILE_INITS=1. Add -ldl in (again) since it _is_ sometimes needed on modern linux.
DETAILS:

Most are explained above, but the -ldl is (almost) always there with gs builds since
it is in the EXTRALIBS list, but main/pcl6_gcc.mak sets EXTRALIBS _before_ the
gs/src/unixgcc.mak runs, so the EXTRALIBS in unixgcc.mak are ignored. This was
needed on my ubuntu gutsy system but not Ralph's ???. It works fine on systems
tested where it was not required.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@2954 06663e23-700e-0410-b217-a244a6096597