Fix dependency problems that caused obj/gs_init.ps to be missed. Also fix a typo that caused CMap files to be missing.
DETAILS:

This method of putting the gs_init.ps dependency down in lib.mak require
us to get gs_init.ps. This hack was discussed with Ralph and is an
alternative to adding a dependency to all of the top level makefiles.
The dummy gs_init.ps is only made when we are NOT including the PS
interpreter (PSI_INCLUDED) since we need the real one for the language
switch build. Note that the dummy file doesn't get added to the %rom%
file system since PS_ROMFS_ARGS is empty if int.mak is not included.

git-svn-id: http://svn.ghostscript.com/ghostpcl/trunk/ghostpcl@2963 06663e23-700e-0410-b217-a244a6096597