Fix (clist interpreter) : Transparency caused a wrong color depth.
DETAILS :

Bug 689520 Regression: 1bpp and grayscale output broken.

The clist interpreter clist_playback_band maintains 2 devices :
the clist reader and target. The clist reader device stores
private interpreter's data. However using a device structure
for this purpose is pretty confusing, because the target
device may have same device fields with different values.
Particularly while processing a transparency, the target device
chage to a transparency buffer device,
which may have a different number of color components.

The problem happens due to the clist interpreter used
a wrong device to obtain the number of color components.
This patch changes the code with taking color components
from the target device.

Now we don't try to analyze all other access to the clist device -
it should be a subject of a separate job.
Therefore the code may need further improvements.
The right way would be to define a special structure for the clist interpreter,
and to drop the clist reader device at all.

EXPECTED DIFFERENCES :

None (comparefiles don't include a test for it; to be added after closing the bug).

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8378 a1074d23-0009-0410-80fe-cf8c14f379e6