Store the posix persistent cache's last modified line as an unsigned long to avoid portability problems when reading and writing. Bug 689604.
DETAILS:

Previously we used time_t directly, but read and wrote it to the
filesystem using the %ld printf format specifier, which is wrong
on systems where long int and time_t are different widths. Instead
we stort it as an unsigned long int, coercing the return value of
time(). This will be a year 2038 problem on systems with 32 bit longs,
but seems the better option for portability now. There is no cast,
so the compiler should warn if this loses precision.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8435 a1074d23-0009-0410-80fe-cf8c14f379e6