Remove the unused composite_rop device.
DETAILS:

This was an attempt to implement PCL raster operations as an interposed
device which first rendered to an intermediate buffer and them
composited with the underlying device buffer, similar to how the
"PDF 1.4 Transparency" device works.

While this would be a nice approach, allowing raster operations to
function in more color spaces, it was never completed, and presents
a maintenance burden for ongoing clist improvements. We are therefore
removing it. It can be reinstated later as a starting point for further
development if needed.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8447 a1074d23-0009-0410-80fe-cf8c14f379e6