Remove the SYSTEM_CONSTANTS_ARE_WRITABLE compile-time define.
Details:

This was added so a particular application could rewrite the product name,
version, etc. after loading the library. We're not aware of anyone still
using the feature, and if so they can just change the source, which is
more appropriate in any case.

Also, remove USE_ASM from the msvc project file.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8471 a1074d23-0009-0410-80fe-cf8c14f379e6