Fix (shadings) : A memory deallocation problem.
DETAILS :

Bug 689660 "memory deallication problems with shadings".

The old code used pdev->memory for temporary storing
an intersection of clipping path with shading bbox.
This is the global memory space.
However if the intersection is trivial,
the path of the clipping path is copied from the
original clipping path, which is allocated in the local memory space.
Later it causes a freeing of path segments to a wrong memory space.

This patch chooses the original clipping path's memory space
when the clipping path is available, and the global space otherwise.

EXPECTED DIFFERENCES :

None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8498 a1074d23-0009-0410-80fe-cf8c14f379e6